requirements.txt
absl-py==0.7.1
appdirs==1.4.3
APScheduler==3.5.3
asn1crypto==0.24.0
astor==0.7.1
baidu-aip==2.2.5.2
beautifulsoup4==4.7.1
bleach==1.5.0
captcha==0.3
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.5
cycler==0.10.0
DBUtils==1.3
easydict==1.8
Flask==1.0.2
Flask-APScheduler==1.11.0
fonttools==3.42.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
html5lib==0.9999999
idna==2.7
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lxml==4.3.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mock==3.0.5
numpy==1.16.2
olefile==0.46
opencv-python==3.4.5.20
pandas==0.23.4
pdfkit==0.6.1
Pillow==6.1.0
protobuf==3.6.1
PyAutoGUI==0.9.42
pycparser==2.19
pyee==6.0.0
PyExecJS==1.5.1
PyGetWindow==0.0.5
pymongo==3.8.0
PyMsgBox==1.0.6
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyPDF2==1.26.0
pyppeteer==0.0.25
PyRect==0.1.4
PyScreeze==0.1.21
python-dateutil==2.8.0
python-pptx==0.6.18
PyTweening==1.0.3
pytz==2018.9
PyYAML==3.13
redis==3.1.0
reportlab==3.5.26
requests==2.19.1
schedule==0.6.0
scikit-learn==0.20.2
scipy==1.2.1
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==1.7.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.36.1
tzlocal==1.5.1
urllib3==1.23
urllib5==5.0.0
webencodings==0.5.1
websockets==8.0.2
Werkzeug==0.15.2
XlsxWriter==1.2.1
xlwt==1.3.0
